Add a function to do some partial reformatting of the slides.
It will call an autoload function defined for each file type, which
formats one line at a time. State information, which may be needed to
format one line based on previous lines, is maintained and passed into
the file-specific autoload function.

To implement this for a filetype, foo for example, create a file named
autoload/foo.vim, and write a function with the signature:
  foo#format(text, state)
Its return value is a 2-item list consisting of:
  1. a list of 1 or more strings to be added to the replacement slide
  2. the new value of state