Specify a custom syntax highlighting to replace the source file's.
The slides produced from markdown files no longer follow markdown syntax
rules. There needs to be some way to specify an alternate syntax file
for these slides. If a ...#set_filetype() function is implemented in the
filetype's autoload module, then it will be used to set the filetype of
the slide show; otherwise, the original source file's type will be used.

This works out nicely because (for now) markdown is the only filetype
that implements this. If a different filetype is used as source for the
slides, it will continue with its current behavior, setting the slide
show's filetype to the same as its source file's.