Fix an annoyance related to using <Esc> and <Left> for navigation.
If you choose to set <Esc> to be the quit key, and <Left> to be the
previous slide key, you will experience very odd behavior related to the
way Vim processes "grey keys". If you press <Left> to go back a slide,
the terminal sends <Esc>[D. When Vim gets this sequence, it sees that
<Esc> has been remapped, and does that (quit the slide show) instead.
Then it continues with [D, and causes all kinds of havoc in the current
buffer. The problem and a solution is described in this StackOverflow
discussion: https://stackoverflow.com/a/20458579/510067