Add support for formatting tables.
Tables are a Github-flavored extension to Markdown. To be used by this
plugin, they must be defined with vertical bars at the beginning and end
of the line. This macro doesn't do any spacing adjustments, except for
centering the whole table, so the user will need to align the columns
manually in the Markdown file.