Prevent vimwiki's overt control of Markdown files when presenting.
If installed, vimwiki will take control of Markdown files, and change
their filetype to vimwiki. When presenting such a file, presenting.vim
will try to find a syntax file named "presenting_vimwiki.vim". It won't
find such a file, and will resort to the original filetype "vimwiki" for
syntax highlighting. This doesn't have all the nice syntax highlighting
features that "presenting_markdown.vim" provides.

So to prevent that hijacking, we just need to look for a filetype of
"vimwiki", and change it to "markdown" when starting the slide show.