Fix off-center tables with better markers in place of TH and TR.
TH and TR were appended to the rows before centering, which caused the
table to be slightly off. Now the table is centered first and then the
rows are prepended with the correct marker for highlighting. The markers
were also switched to «th» and «tr» to prevent any accidental hiding of
legitimate text, such as "The" or "Trumpet", which could be the
beginning of a line of text.